+++ src/components/main/main.jsx
@@ -1,7 +1,8 @@
-import React, {Component} from 'react';
+import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
+import PlacesList from './../places-list/places-list.jsx';
 
-export default class Main extends Component {
+export default class Main extends PureComponent {
   render() {
     return <div className="page page--gray page--main" key="app-main">
       <header className="header">
@@ -85,9 +86,7 @@ export default class Main extends Component {
                   <li className="places__option" tabIndex="0">Top rated first</li>
                 </ul>
               </form>
-              <div className="cities__places-list places__list tabs__content">
-                {this.properties}
-              </div>
+              <PlacesList key="PlacesList" properties={this.props.properties} onClick={this.props.onClick}/>
             </section>
             <div className="cities__right-section">
               <section className="cities__map map"></section>
@@ -96,42 +95,6 @@ export default class Main extends Component {
         </div>
       </main>
     </div>;
-  }
-
-  get properties() {
-    return this.props.properties.map((item) => (
-      <article className="cities__place-card place-card" key={item.id}>
-        <div className="cities__image-wrapper place-card__image-wrapper">
-          <a href="#">
-            <img className="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image"></img>
-          </a>
-        </div>
-        <div className="place-card__info">
-          <div className="place-card__price-wrapper">
-            <div className="place-card__price">
-              <b className="place-card__price-value">&euro;132</b>
-              <span className="place-card__price-text">&#47;&nbsp;night</span>
-            </div>
-            <button className="place-card__bookmark-button button" type="button">
-              <svg className="place-card__bookmark-icon" width="18" height="19">
-                <use xlinkHref="#icon-bookmark"></use>
-              </svg>
-              <span className="visually-hidden">To bookmarks</span>
-            </button>
-          </div>
-          <div className="place-card__rating rating">
-            <div className="place-card__stars rating__stars">
-              <span style={{width: `80%`}}></span>
-              <span className="visually-hidden">Rating</span>
-            </div>
-          </div>
-          <h2 className="place-card__name">
-            <a href="#" onClick={this.props.onClick}>{item.caption}</a>
-          </h2>
-          <p className="place-card__type">Apartment</p>
-        </div>
-      </article>
-    ));
   }
 }
