+++ src/mocks/offers.js
@@ -1,0 +1,20 @@
+const offers = [
+  {
+    id: `prop-1`,
+    caption: `Beautiful & luxurious apartment at great location`,
+  },
+  {
+    id: `prop-2`,
+    caption: `Wood and stone place`,
+  },
+  {
+    id: `prop-3`,
+    caption: `Canal View Prinsengracht`,
+  },
+  {
+    id: `prop-4`,
+    caption: `Nice, cozy, warm big bed apartment`,
+  },
+];
+
+export default offers;
